requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0